detect/http_host: move tests into tests/
